rename package pkg/datarepair to satellite/repair
