+++ satellite/repair/checker/checker_test.go
@@ -16,9 +16,9 @@ import (
 	"storj.io/storj/internal/testcontext"
 	"storj.io/storj/internal/testplanet"
 	"storj.io/storj/internal/teststorj"
-	"storj.io/storj/pkg/datarepair/checker"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
+	"storj.io/storj/satellite/repair/checker"
 	"storj.io/storj/storage"
 )
