+++ satellite/repair/irreparable/inspector.go
